+++ packages/test/src/test-integration-split-two.ts
@@ -15,7 +15,7 @@ import { msToNumber, tsToMs } from '@temporalio/common/lib/time';
 import { decode as payloadDecode, decodeFromPayloadsAtIndex } from '@temporalio/common/lib/internal-non-workflow';
 
 import { condition, defineQuery, setHandler, sleep } from '@temporalio/workflow';
-import { configurableHelpers } from './helpers-integration';
+import { configurableHelpers, createTestWorkflowBundle } from './helpers-integration';
 import * as activities from './activities';
 import * as workflows from './workflows';
 import { makeTestFn, configMacro } from './configured-integration-helpers';
@@ -24,7 +24,7 @@ import { makeTestFn, configMacro } from './configured-integration-helpers';
 //  - review the files where these workflows are shared
 export * from './workflows';
 
-const test = makeTestFn({ workflowsPath: __filename });
+const test = makeTestFn(() => createTestWorkflowBundle({ workflowsPath: __filename }));
 test.macro(configMacro);
 
 test('WorkflowOptions are passed correctly with defaults', configMacro, async (t, config) => {
@@ -677,123 +677,6 @@ test(
     t.pass();
   }
 );
-
-// TODO(thomas): fix
-/*
-  if ('promiseHooks' in v8) {
-    // Skip in old node versions
-    test('Stack trace query returns stack that makes sense', configMacro, async (t, config) => {
-      const { env, createWorkerWithDefaults } = config;
-  
-      const { executeWorkflow, createWorker } = configurableHelpers(t, t.context.workflowBundle, env);
-      const worker = await createWorkerWithDefaults(t, { activities });
-      const rawStacks = await worker.runUntil(executeWorkflow(workflows.stackTracer));
-  
-      const [stack1, stack2] = rawStacks.map((r) =>
-        r
-          .split('\n\n')
-          .map((s) => cleanOptionalStackTrace(`\n${s}`))
-          .join('\n')
-      );
-      // Can't get the Trigger stack cleaned, this is okay for now
-      // NOTE: we check endsWith because under certain conditions we might see Promise.race in the trace
-      t.true(
-        stack1.endsWith(
-          `
-    at Function.all (<anonymous>)
-    at stackTracer (test/src/workflows/stack-tracer.ts)
-  
-    at stackTracer (test/src/workflows/stack-tracer.ts)
-  
-    at Promise.then (<anonymous>)
-    at Trigger.then (workflow/src/trigger.ts)`
-        ),
-        `Got invalid stack:\n--- clean ---\n${stack1}\n--- raw ---\n${rawStacks[0]}`
-      );
-      t.is(
-        stack2,
-        `
-    at executeChild (workflow/src/workflow.ts)
-    at stackTracer (test/src/workflows/stack-tracer.ts)
-  
-    at new Promise (<anonymous>)
-    at timerNextHandler (workflow/src/workflow.ts)
-    at sleep (workflow/src/workflow.ts)
-    at stackTracer (test/src/workflows/stack-tracer.ts)
-  
-    at stackTracer (test/src/workflows/stack-tracer.ts)`
-      );
-    });
-    
-    
-    test('Enhanced stack trace returns trace that makes sense', configMacro, async (t, config) => {
-      const { env, createWorkerWithDefaults } = config;
-  
-      const { executeWorkflow, createWorker } = configurableHelpers(t, t.context.workflowBundle, env);
-      const worker = await createWorkerWithDefaults(t, { activities });
-      const enhancedStack = await worker.runUntil(executeWorkflow(workflows.enhancedStackTracer));
-  
-      const stacks = enhancedStack.stacks.map((s) => ({
-        locations: s.locations.map((l) => ({
-          ...l,
-          ...(l.file_path
-            ? { file_path: l.file_path.replace(path.resolve(__dirname, '../../../'), '').replace(/\\/g, '/') }
-            : undefined),
-        })),
-      }));
-      t.is(enhancedStack.sdk.name, 'typescript');
-      t.is(enhancedStack.sdk.version, pkg.version); // Expect workflow and worker versions to match
-      t.deepEqual(stacks, [
-        {
-          locations: [
-            {
-              function_name: 'Function.all',
-              internal_code: false,
-            },
-            {
-              file_path: '/packages/test/src/workflows/stack-tracer.ts',
-              function_name: 'enhancedStackTracer',
-              line: 32,
-              column: 35,
-              internal_code: false,
-            },
-          ],
-        },
-        {
-          locations: [
-            {
-              file_path: '/packages/test/src/workflows/stack-tracer.ts',
-              function_name: 'enhancedStackTracer',
-              line: 32,
-              column: 35,
-              internal_code: false,
-            },
-          ],
-        },
-        {
-          locations: [
-            {
-              function_name: 'Promise.then',
-              internal_code: false,
-            },
-            {
-              file_path: '/packages/workflow/src/trigger.ts',
-              function_name: 'Trigger.then',
-              line: 47,
-              column: 24,
-              internal_code: false,
-            },
-          ],
-        },
-      ]);
-      const expectedSources = ['../src/workflows/stack-tracer.ts', '../../workflow/src/trigger.ts'].map((p) => [
-        path.resolve(__dirname, p),
-        [{ content: readFileSync(path.resolve(__dirname, p), 'utf8'), line_offset: 0 }],
-      ]);
-      t.deepEqual(Object.entries(enhancedStack.sources), expectedSources);
-    });
-  }
-  */
 
 const mutateWorkflowStateQuery = defineQuery<void>('mutateWorkflowState');
 export async function queryAndCondition(): Promise<void> {
